% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/andthis-package.R
\docType{package}
\name{andthis-package}
\alias{andthis}
\alias{andthis-package}
\title{andthis: Jon's Convenience Functions}
\description{
My rprofile was getting too big, so I created a metapackage, finally. This also allows me to create templates for usethis.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/jonthegeek/andthis}
  \item \url{https://jonthegeek.github.io/andthis/}
  \item Report bugs at \url{https://github.com/jonthegeek/andthis/issues}
}

}
\author{
\strong{Maintainer}: Jon Harmon \email{jonthegeek@gmail.com} (\href{https://orcid.org/0000-0003-4781-4346}{ORCID})

}
\keyword{internal}
