% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_here}
\alias{browse_here}
\title{Open this project's root directory}
\usage{
browse_here()
}
\value{
The root directory (invisibly).
}
\description{
Open the root directory of the current project in the system's file browser.
}
