% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_repo}
\alias{browse_repo}
\title{Open this project's GitHub repo in browser}
\usage{
browse_repo(path = ".")
}
\arguments{
\item{path}{Path to the root of the project.}
}
\value{
The URL of the GitHub repository (invisibly).
}
\description{
Open the remote GitHub repository for the project at \code{path} in the system's
web browser.
}
