% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browse.R
\name{browse_temp}
\alias{browse_temp}
\title{Open the current tempdir}
\usage{
browse_temp()
}
\value{
The path to the tempdir (invisibly).
}
\description{
Open the current R session's tempdir in the system's file browser.
}
