% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-loading.R
\name{club}
\alias{club}
\title{Open the RStudio project for a book club}
\usage{
club(club, new_session = TRUE)
}
\arguments{
\item{club}{The abbreviation for the club.}

\item{new_session}{Whether to open the project in a new RStudio session.}
}
\value{
Called for side effects. Probably something invisibly.
}
\description{
Open the RStudio project for a book club
}
