% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ds.R
\name{create_club}
\alias{create_club}
\title{Create an R4DS book club}
\usage{
create_club(book_abbr)
}
\arguments{
\item{book_abbr}{The short name for the book. Should match the channel.}
}
\value{
The path to the new project, invisibly.
}
\description{
Create an R4DS book club
}
