% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r4ds.R
\name{create_club_chapter_rmds}
\alias{create_club_chapter_rmds}
\title{Copy File Template for Chapters}
\usage{
create_club_chapter_rmds(chapters)
}
\arguments{
\item{chapters}{A character vector of chapter titles. For now the assumption
is that the vector contains numbered chapters starting with chapter 1.}
}
\value{
NULL (invisibly)
}
\description{
Create a separate Rmd for each chapter, and apply the basic template
information.
}
