% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{.extract_precommit_blocks}
\alias{.extract_precommit_blocks}
\title{Parse formatted hooks into a list of hooks}
\usage{
.extract_precommit_blocks(existing_pre_commit)
}
\arguments{
\item{existing_pre_commit}{A character vector with one element per row of the
existing pre-commit hook, minus the boilerplate at the start.}
}
\value{
A list of hooks.
}
\description{
Parse formatted hooks into a list of hooks
}
\keyword{internal}
