% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{.finalize_hooks}
\alias{.finalize_hooks}
\title{Sort hooks and combine}
\usage{
.finalize_hooks(prepared_hooks)
}
\arguments{
\item{prepared_hooks}{A list of hooks.}
}
\value{
A character vector with the hook contents, in priority order.
}
\description{
Sort hooks and combine
}
\keyword{internal}
