% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{.parse_existing_pre_commit}
\alias{.parse_existing_pre_commit}
\title{Load existing pre-commit hooks and prep for recombining}
\usage{
.parse_existing_pre_commit()
}
\value{
A list of hooks.
}
\description{
Load existing pre-commit hooks and prep for recombining
}
\keyword{internal}
