% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{.prepare_existing_hooks}
\alias{.prepare_existing_hooks}
\title{Prep existing hooks for recombining}
\usage{
.prepare_existing_hooks(existing_pre_commit)
}
\arguments{
\item{existing_pre_commit}{A character vector with one element per row of the
existing pre-commit hook, minus the boilerplate at the start.}
}
\value{
A list of hooks.
}
\description{
Prep existing hooks for recombining
}
\keyword{internal}
