% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{.prepare_hook}
\alias{.prepare_hook}
\title{Clean up a hook and prep it for use}
\usage{
.prepare_hook(hook, priority = 0)
}
\arguments{
\item{hook}{The hook as a character vector, with one line per element of the
vector.}

\item{priority}{The integer priority of the hook. Lower-priority hooks will
be placed later in the script.}
}
\value{
A character vector with class "hook," with the priority as an
attribute.
}
\description{
Clean up a hook and prep it for use
}
\keyword{internal}
