% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{finish_package_setup}
\alias{finish_package_setup}
\title{Run the Rest of the Things}
\usage{
finish_package_setup(r4ds = TRUE)
}
\arguments{
\item{r4ds}{Whether to create the package under the r4ds organization
(default \code{TRUE}).}
}
\value{
The path to the package, probably.
}
\description{
This must execute in a package directory. It assumes a lot.
}
