% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{install_this}
\alias{install_this}
\title{Install the current repo}
\usage{
install_this()
}
\value{
The result of the installation, invisibly.
}
\description{
Install from GitHub to make sure things like renv understand where this
package is coming from.
}
