% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/project-loading.R
\name{proj}
\alias{proj}
\title{Open an RStudio project}
\usage{
proj(project, new_session = TRUE)
}
\arguments{
\item{project}{The name of an existing project.}

\item{new_session}{Whether to open the project in a new RStudio session.}
}
\value{
Called for side effects. Probably something invisibly.
}
\description{
Open an RStudio project by name. If the project does not exist, this function
will throw an error.
}
