% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{protect_main}
\alias{protect_main}
\title{Stop yourself from checking into the main branch}
\usage{
protect_main()
}
\value{
\code{NULL} (invisibly).
}
\description{
Stop yourself from checking into the main branch
}
