% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_tree}
\alias{repo_tree}
\title{Wrapper of gh_tree_remote for testing}
\usage{
repo_tree(path = ".")
}
\arguments{
\item{path}{Path that is contained within a git repo.}
}
\value{
If the repo has a github remote, a list containing \code{username}
and \code{repo}. Otherwise, an error.
}
\description{
Wrapper of gh_tree_remote for testing
}
\keyword{internal}
