% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repo.R
\name{repo_url}
\alias{repo_url}
\title{Return the URL for a GitHub repo}
\usage{
repo_url(path = ".")
}
\arguments{
\item{path}{Path to the root of a project that has a GitHub remote.}
}
\value{
The URL of the GitHub repository.
}
\description{
Return the URL for the GitHub repository for the project at \code{path}.
}
