% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{use_precommit_hook}
\alias{use_precommit_hook}
\title{Add something to the precommit hook}
\usage{
use_precommit_hook(hook)
}
\arguments{
\item{hook}{Character. The hook to add, with one line of output per element
of the vector.}
}
\value{
\code{NULL} (invisibly).
}
\description{
Unlike \code{\link[usethis:use_git_hook]{usethis::use_git_hook()}}, this function checks to see if you already
have a pre-commit hook, and attempts to add to it while avoiding duplication.
The checks are imperfect, which is probably why this isn't in usethis this
way.
}
