% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{create_package}
\alias{create_package}
\title{Create a Jon Package}
\usage{
create_package(name, ...)
}
\arguments{
\item{name}{The name of the package}

\item{...}{Additional parameters passed on to \code{\link[usethis:create_package]{usethis::create_package()}}.}
}
\value{
Path to the newly created project or package, invisibly.
}
\description{
Create a Jon Package
}
