% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package-setup.R
\name{use_precommit_hook}
\alias{use_precommit_hook}
\title{Add something to the precommit hook}
\usage{
use_precommit_hook(hook)
}
\value{
\code{NULL} (invisibly).
}
\description{
Add something to the precommit hook
}
